feat(parser): split statement mods, linter fixes
