+++ gen/bobgen-psql/driver/parser/mods.go
@@ -2,10 +2,8 @@ package parser
 
 import (
 	"fmt"
-	"strconv"
 
 	pg "github.com/pganalyze/pg_query_go/v6"
-	"github.com/stephenafamo/bob/clause"
 	"github.com/stephenafamo/bob/internal"
 )
 
@@ -33,674 +31,4 @@ func (w *walker) modWithClause(with *pg.WithClause, info nodeInfo) {
 			)...,
 		)
 	}
-}
-
-//nolint:gocyclo,maintidx
-func (w *walker) modSelectStatement(stmt *pg.Node_SelectStmt, info nodeInfo) {
-	if withInfo, ok := info.children["WithClause"]; ok {
-		w.modWithClause(stmt.SelectStmt.WithClause, withInfo)
-	}
-
-	main := stmt.SelectStmt
-	mainInfo := info
-
-	var combines []combine
-
-	for main.Larg != nil {
-		combines = append(combines, combine{
-			Operation:  main.Op,
-			All:        main.All,
-			SelectStmt: main.Rarg,
-			Info:       mainInfo.children["Rarg"],
-		})
-		main = main.Larg
-		mainInfo = mainInfo.children["Larg"]
-	}
-
-	if distinct := main.DistinctClause; distinct != nil {
-		if len(distinct) == 1 && distinct[0].Node == nil {
-			fmt.Fprintln(w.mods, "q.Distinct.On = []any{}")
-		} else {
-			distinctInfo := info.children["DistinctClause"]
-			w.editRules = append(w.editRules,
-				internal.RecordPoints(
-					int(distinctInfo.start),
-					int(distinctInfo.end-1),
-					func(start, end int) error {
-						fmt.Fprintf(
-							w.mods,
-							"q.Distinct.On = append(q.Distinct.On,EXPR.subExpr(%d, %d))\n",
-							start, end,
-						)
-						return nil
-					},
-				)...,
-			)
-		}
-	}
-
-	if targetInfo, ok := mainInfo.children["TargetList"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(targetInfo.start),
-				int(targetInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendSelect(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if fromInfo, ok := mainInfo.children["FromClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(fromInfo.start),
-				int(fromInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.SetTable(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if whereInfo, ok := mainInfo.children["WhereClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(whereInfo.start),
-				int(whereInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendWhere(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if groupByInfo, ok := mainInfo.children["GroupClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(groupByInfo.start),
-				int(groupByInfo.end)-1,
-				func(start, end int) error {
-					if main.GroupDistinct {
-						fmt.Fprint(w.mods, "q.SetGroupByDistinct(true)\n")
-					}
-					fmt.Fprintf(w.mods, "q.AppendGroup(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if havingInfo, ok := mainInfo.children["HavingClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(havingInfo.start),
-				int(havingInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendHaving(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if windowsInfo, ok := mainInfo.children["WindowClause"]; ok {
-		for _, windowInfo := range windowsInfo.children {
-			nameStart := w.getStartOfTokenBefore(windowInfo.start, pg.Token_IDENT)
-			w.editRules = append(w.editRules,
-				internal.RecordPoints(
-					int(nameStart),
-					int(windowInfo.end)-1,
-					func(start, end int) error {
-						fmt.Fprintf(w.mods, "q.AppendWindow(EXPR.subExpr(%d, %d))\n", start, end)
-						return nil
-					},
-				)...,
-			)
-		}
-	}
-
-	if len(combines) > 0 {
-		w.imports = append(w.imports, []string{"github.com/stephenafamo/bob/clause"})
-	}
-	for i := len(combines) - 1; i >= 0; i-- {
-		combine := combines[i]
-
-		strategy := ""
-		switch combine.Operation {
-		case pg.SetOperation_SETOP_UNION:
-			strategy = "UNION"
-		case pg.SetOperation_SETOP_INTERSECT:
-			strategy = "INTERSECT"
-		case pg.SetOperation_SETOP_EXCEPT:
-			strategy = "EXCEPT"
-		}
-
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(combine.Info.start),
-				int(combine.Info.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, `
-                        q.AppendCombine(clause.Combine{
-                            Strategy: "%s",
-                            All: %t,
-                            Query: bob.BaseQuery[bob.Expression]{
-                                Expression: EXPR.subExpr(%d, %d),
-                                QueryType: bob.QueryTypeSelect,
-                                Dialect: dialect.Dialect,
-                            },
-                        })
-                    `, strategy, combine.All, start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if limitInfo, ok := info.children["LimitCount"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(limitInfo.start),
-				int(limitInfo.end)-1,
-				func(start, end int) error {
-					switch stmt.SelectStmt.LimitOption {
-					case pg.LimitOption_LIMIT_OPTION_COUNT:
-						fmt.Fprintf(w.mods, "q.CombinedLimit.SetLimit(EXPR.subExpr(%d, %d))\n", start, end)
-					case pg.LimitOption_LIMIT_OPTION_WITH_TIES:
-						w.imports = append(w.imports, []string{"github.com/stephenafamo/bob/clause"})
-						fmt.Fprintf(w.mods, `q.CombinedFetch.SetFetch(clause.Fetch{
-								Count: EXPR.subExpr(%d, %d),
-								WithTies: true,
-							})
-						`, start, end)
-					}
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if offsetInfo, ok := mainInfo.children["LimitOffset"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(offsetInfo.start),
-				int(offsetInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.CombinedOffset.SetOffset(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if orderInfo, ok := info.children["SortClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(orderInfo.start),
-				int(orderInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.CombinedOrder.AppendOrder(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	for i, lockClause := range stmt.SelectStmt.LockingClause {
-		lock, ok := lockClause.Node.(*pg.Node_LockingClause)
-		if !ok || lock.LockingClause.Strength < 2 {
-			continue
-		}
-
-		var bobLock clause.Lock
-
-		switch lock.LockingClause.Strength {
-		case pg.LockClauseStrength_LCS_FORKEYSHARE:
-			bobLock.Strength = "KEY SHARE"
-		case pg.LockClauseStrength_LCS_FORSHARE:
-			bobLock.Strength = "SHARE"
-		case pg.LockClauseStrength_LCS_FORNOKEYUPDATE:
-			bobLock.Strength = "NO KEY UPDATE"
-		case pg.LockClauseStrength_LCS_FORUPDATE:
-			bobLock.Strength = "UPDATE"
-		}
-
-		switch lock.LockingClause.WaitPolicy {
-		case pg.LockWaitPolicy_LockWaitSkip:
-			bobLock.Wait = "SKIP LOCKED"
-		case pg.LockWaitPolicy_LockWaitError:
-			bobLock.Wait = "NOWAIT"
-		}
-
-		if len(lock.LockingClause.LockedRels) > 0 {
-			lockInfo := info.children["LockingClause"].children[strconv.Itoa(i)]
-			bobLock.Tables = []string{w.input[lockInfo.start:lockInfo.end]}
-		}
-
-		w.imports = append(w.imports, []string{"github.com/stephenafamo/bob/clause"})
-		fmt.Fprintf(w.mods, "q.AppendLock(%#v)\n", bobLock)
-	}
-}
-
-func (w *walker) modInsertStatement(stmt *pg.Node_InsertStmt, info nodeInfo) {
-	if withInfo, ok := info.children["WithClause"]; ok {
-		w.modWithClause(stmt.InsertStmt.WithClause, withInfo)
-	}
-
-	if intoInfo, ok := info.children["Relation"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(intoInfo.start),
-				int(intoInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.TableRef.Expression = EXPR.subExpr(%d, %d)\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	colNames := make([]string, len(stmt.InsertStmt.Cols))
-	for i := range stmt.InsertStmt.Cols {
-		colNameInfo := info.children["Cols"].children[strconv.Itoa(i)]
-		colNames[i] = w.names[colNameInfo.position()]
-	}
-	if len(colNames) > 0 {
-		fmt.Fprintf(w.mods, "q.TableRef.Columns = %#v\n", colNames)
-	}
-
-	if selectInfo, ok := info.children["SelectStmt"].children["SelectStmt"]; ok {
-		w.editRules = append(w.editRules, internal.RecordPoints(
-			int(selectInfo.start), int(selectInfo.end)-1,
-			func(start, end int) error {
-				fmt.Fprintf(w.mods, `q.Query = bob.BaseQuery[bob.Expression]{
-						Expression: EXPR.subExpr(%d, %d),
-						Dialect: dialect.Dialect,
-						QueryType: bob.QueryTypeSelect,
-						}
-					`, start, end)
-				return nil
-			},
-		)...)
-	}
-
-	switch stmt.InsertStmt.Override {
-	case pg.OverridingKind_OVERRIDING_USER_VALUE:
-		fmt.Fprintln(w.mods, `q.Overriding = "USER"`)
-	case pg.OverridingKind_OVERRIDING_SYSTEM_VALUE:
-		fmt.Fprintln(w.mods, `q.Overriding = "SYSTEM"`)
-	}
-
-	if conflictInfo, ok := info.children["OnConflictClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(conflictInfo.start),
-				int(conflictInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.SetConflict(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if returnInfo, ok := info.children["ReturningList"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(returnInfo.start),
-				int(returnInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendReturning(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-}
-
-func (w *walker) modUpdateStatement(stmt *pg.Node_UpdateStmt, info nodeInfo) {
-	if withInfo, ok := info.children["WithClause"]; ok {
-		w.modWithClause(stmt.UpdateStmt.WithClause, withInfo)
-	}
-
-	if tableInfo, ok := info.children["Relation"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(tableInfo.start),
-				int(tableInfo.end)-1,
-				func(start, end int) error {
-					if !stmt.UpdateStmt.GetRelation().GetInh() {
-						fmt.Fprintln(w.mods, "q.Only = true")
-					}
-					fmt.Fprintf(w.mods, "q.Table.Expression = EXPR.subExpr(%d, %d)\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if targetInfo, ok := info.children["TargetList"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(targetInfo.start),
-				int(targetInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendSet(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if fromInfo, ok := info.children["FromClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(fromInfo.start),
-				int(fromInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.SetTable(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if whereInfo, ok := info.children["WhereClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(whereInfo.start),
-				int(whereInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendWhere(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if returnInfo, ok := info.children["ReturningList"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(returnInfo.start),
-				int(returnInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendReturning(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-}
-
-func (w *walker) modDeleteStatement(stmt *pg.Node_DeleteStmt, info nodeInfo) {
-	if withInfo, ok := info.children["WithClause"]; ok {
-		w.modWithClause(stmt.DeleteStmt.WithClause, withInfo)
-	}
-
-	if tableInfo, ok := info.children["Relation"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(tableInfo.start),
-				int(tableInfo.end)-1,
-				func(start, end int) error {
-					if !stmt.DeleteStmt.GetRelation().GetInh() {
-						fmt.Fprintln(w.mods, "q.Only = true")
-					}
-					fmt.Fprintf(w.mods, "q.Table.Expression = EXPR.subExpr(%d, %d)\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if usingInfo, ok := info.children["UsingClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(usingInfo.start),
-				int(usingInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.SetTable(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if whereInfo, ok := info.children["WhereClause"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(whereInfo.start),
-				int(whereInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendWhere(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if returnInfo, ok := info.children["ReturningList"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(returnInfo.start),
-				int(returnInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendReturning(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-}
-
-func (w *walker) modMergeStatement(stmt *pg.Node_MergeStmt, info nodeInfo) {
-	if withInfo, ok := info.children["WithClause"]; ok {
-		w.modWithClause(stmt.MergeStmt.WithClause, withInfo)
-	}
-
-	if tableInfo, ok := info.children["Relation"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(tableInfo.start),
-				int(tableInfo.end)-1,
-				func(start, end int) error {
-					if !stmt.MergeStmt.GetRelation().GetInh() {
-						fmt.Fprintln(w.mods, "q.Only = true")
-					}
-					fmt.Fprintf(w.mods, "q.Table.Expression = EXPR.subExpr(%d, %d)\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if sourceInfo, ok := info.children["SourceRelation"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(sourceInfo.start),
-				int(sourceInfo.end)-1,
-				func(start, end int) error {
-					if src := stmt.MergeStmt.GetSourceRelation(); src != nil {
-						if rangeVar, ok := src.Node.(*pg.Node_RangeVar); ok && rangeVar.RangeVar != nil && !rangeVar.RangeVar.GetInh() {
-							fmt.Fprintln(w.mods, "q.Using.Only = true")
-						}
-					}
-					fmt.Fprintf(w.mods, "q.Using.Source = EXPR.subExpr(%d, %d)\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	if onInfo, ok := info.children["JoinCondition"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(onInfo.start),
-				int(onInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.Using.Condition = EXPR.subExpr(%d, %d)\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-
-	whenInfos := info.children["MergeWhenClauses"]
-	whenCount := 0
-	for i, whenNode := range stmt.MergeStmt.MergeWhenClauses {
-		when := whenNode.GetMergeWhenClause()
-		if when == nil {
-			continue
-		}
-
-		whenType, ok := mergeMatchKindToDialectWhenType(when.MatchKind)
-		if !ok {
-			w.errors = append(w.errors, fmt.Errorf("unsupported MERGE match kind: %s", when.MatchKind.String()))
-			continue
-		}
-
-		actionType, ok := mergeCommandTypeToDialectActionType(when.CommandType)
-		if !ok {
-			w.errors = append(w.errors, fmt.Errorf("unsupported MERGE command type: %s", when.CommandType.String()))
-			continue
-		}
-
-		fmt.Fprintf(w.mods,
-			"q.When = append(q.When, dialect.MergeWhen{Type: %s, Action: dialect.MergeAction{Type: %s}})\n",
-			whenType,
-			actionType,
-		)
-		currentWhenIndex := whenCount
-		whenCount++
-
-		if overriding, ok := mergeOverridingToDialect(when.Override); ok {
-			fmt.Fprintf(w.mods, "q.When[%d].Action.Overriding = %s\n", currentWhenIndex, overriding)
-		}
-
-		if when.CommandType == pg.CmdType_CMD_INSERT {
-			if cols := mergeInsertColumns(when); len(cols) > 0 {
-				fmt.Fprintf(w.mods, "q.When[%d].Action.Columns = %#v\n", currentWhenIndex, cols)
-			}
-		}
-
-		whenInfo, hasWhenInfo := whenInfos.children[strconv.Itoa(i)]
-		if wrappedInfo, ok := whenInfo.children["MergeWhenClause"]; ok {
-			whenInfo = wrappedInfo
-		}
-
-		if hasWhenInfo {
-			if conditionInfo, ok := whenInfo.children["Condition"]; ok {
-				w.editRules = append(w.editRules,
-					internal.RecordPoints(
-						int(conditionInfo.start),
-						int(conditionInfo.end)-1,
-						func(start, end int) error {
-							fmt.Fprintf(w.mods, "q.When[%d].Condition = EXPR.subExpr(%d, %d)\n", currentWhenIndex, start, end)
-							return nil
-						},
-					)...,
-				)
-			}
-
-			switch when.CommandType {
-			case pg.CmdType_CMD_INSERT:
-				if valuesInfo, ok := whenInfo.children["Values"]; ok {
-					w.editRules = append(w.editRules,
-						internal.RecordPoints(
-							int(valuesInfo.start),
-							int(valuesInfo.end)-1,
-							func(start, end int) error {
-								fmt.Fprintf(w.mods, "q.When[%d].Action.Values = append(q.When[%d].Action.Values, EXPR.subExpr(%d, %d))\n", currentWhenIndex, currentWhenIndex, start, end)
-								return nil
-							},
-						)...,
-					)
-				}
-
-			case pg.CmdType_CMD_UPDATE:
-				if setInfo, ok := whenInfo.children["TargetList"]; ok {
-					w.editRules = append(w.editRules,
-						internal.RecordPoints(
-							int(setInfo.start),
-							int(setInfo.end)-1,
-							func(start, end int) error {
-								fmt.Fprintf(w.mods, "q.When[%d].Action.Set = append(q.When[%d].Action.Set, EXPR.subExpr(%d, %d))\n", currentWhenIndex, currentWhenIndex, start, end)
-								return nil
-							},
-						)...,
-					)
-				}
-			}
-		}
-	}
-
-	if returnInfo, ok := info.children["ReturningList"]; ok {
-		w.editRules = append(w.editRules,
-			internal.RecordPoints(
-				int(returnInfo.start),
-				int(returnInfo.end)-1,
-				func(start, end int) error {
-					fmt.Fprintf(w.mods, "q.AppendReturning(EXPR.subExpr(%d, %d))\n", start, end)
-					return nil
-				},
-			)...,
-		)
-	}
-}
-
-func mergeMatchKindToDialectWhenType(kind pg.MergeMatchKind) (string, bool) {
-	switch kind {
-	case pg.MergeMatchKind_MERGE_WHEN_MATCHED:
-		return "dialect.MergeWhenMatched", true
-	case pg.MergeMatchKind_MERGE_WHEN_NOT_MATCHED_BY_SOURCE:
-		return "dialect.MergeWhenNotMatchedBySource", true
-	case pg.MergeMatchKind_MERGE_WHEN_NOT_MATCHED_BY_TARGET:
-		return "dialect.MergeWhenNotMatched", true
-	default:
-		return "", false
-	}
-}
-
-func mergeCommandTypeToDialectActionType(cmd pg.CmdType) (string, bool) {
-	switch cmd {
-	case pg.CmdType_CMD_NOTHING:
-		return "dialect.MergeActionDoNothing", true
-	case pg.CmdType_CMD_DELETE:
-		return "dialect.MergeActionDelete", true
-	case pg.CmdType_CMD_INSERT:
-		return "dialect.MergeActionInsert", true
-	case pg.CmdType_CMD_UPDATE:
-		return "dialect.MergeActionUpdate", true
-	default:
-		return "", false
-	}
-}
-
-func mergeOverridingToDialect(override pg.OverridingKind) (string, bool) {
-	switch override {
-	case pg.OverridingKind_OVERRIDING_USER_VALUE:
-		return "dialect.OverridingUser", true
-	case pg.OverridingKind_OVERRIDING_SYSTEM_VALUE:
-		return "dialect.OverridingSystem", true
-	default:
-		return "", false
-	}
-}
-
-func mergeInsertColumns(when *pg.MergeWhenClause) []string {
-	cols := make([]string, 0, len(when.TargetList))
-	for _, target := range when.TargetList {
-		resTarget := target.GetResTarget()
-		if resTarget == nil || resTarget.Name == "" {
-			continue
-		}
-
-		cols = append(cols, resTarget.Name)
-	}
-
-	return cols
 }
